Add preliminary version of the “partial” function
This is *preliminary* since, despite a lot of effort, the function
doesn’t actually work in a lot of interesting cases (which are
documented). Hadley notes the same in his implementation in
`pryr::partial`, and opts for a much simpler variant that just works.
However, the current implementation has some desirable features which
the simple implementation doesn’t have. The current implementation was
actually preceded by a simple implementation in the spirit of
`pryr::partial`, that was found insufficient in practice.